# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Product Foundation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/account/deletion": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Schedule deleting my account",
        "description": "Type your email to confirm. You can cancel until the date shown.",
        "operationId": "schedule_account_deletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionStatus"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Cancel deleting my account",
        "description": "Keep the account.",
        "operationId": "cancel_account_deletion",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionStatus"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/exports": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Export my data (ZIP)",
        "description": "Everything we store about you. Follow the job, then download the ZIP.",
        "operationId": "export_account",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRead"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "My exports",
        "description": "Your exports, plus this workspace's exports if you may export it.",
        "operationId": "list_exports",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportList"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{export_id}/download": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Download an export",
        "description": "The ZIP file. Only for the person (or workspace admins) it belongs to.",
        "operationId": "download_export",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The ZIP file",
            "content": {
              "application/zip": {}
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeletionRequest": {
        "properties": {
          "confirm": {
            "type": "string",
            "maxLength": 320,
            "minLength": 1,
            "title": "Confirm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirm"
        ],
        "title": "DeletionRequest",
        "description": "Type the confirmation text exactly (the email address, or the workspace name)."
      },
      "DeletionStatus": {
        "properties": {
          "deletion_scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletion Scheduled At"
          }
        },
        "type": "object",
        "required": [
          "deletion_scheduled_at"
        ],
        "title": "DeletionStatus",
        "description": "When the deletion will happen (null = not scheduled)."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "description": "What went wrong."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Error envelope: `{\"error\": {...}}`."
      },
...
      "ExportList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExportRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExportList",
        "description": "Your recent exports (and the workspace's, if you may see them). Newest first."
      },
      "ExportRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ExportScope"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id",
            "description": "Follow progress with GET /api/jobs/{id}."
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "job_id",
          "filename",
          "ready",
          "size_bytes",
          "created_at",
          "expires_at",
          "download_url"
        ],
        "title": "ExportRead",
        "description": "A data export. Download it when `ready` is true."
      },
      "ExportScope": {
        "type": "string",
        "enum": [
          "account",
          "organization"
        ],
        "title": "ExportScope",
        "description": "Whose data is in the export."
      },
...
    }
  }
}